Features

+++ src/main/java/com/magenta/crud/user/UserServiceImpl.java
@@ -2,12 +2,20 @@ package com.magenta.crud.user;
 
 import com.magenta.crud.contract.ContractService;
 import com.magenta.crud.contract.dto.ContractDto;
+import com.magenta.crud.global.dto.ChangeStatusDto;
 import com.magenta.crud.type.Role;
 import com.magenta.crud.type.Status;
+import com.magenta.crud.user.dto.AddFundsDto;
+import com.magenta.crud.user.dto.EditUserDto;
 import com.magenta.crud.user.dto.NewUserDto;
 import com.magenta.crud.user.dto.UserDto;
+import com.magenta.mapper.MyModelMapper;
+import com.magenta.myexception.AuthorizationException;
+import com.magenta.myexception.DatabaseException;
 import com.magenta.myexception.MyException;
-import org.modelmapper.ModelMapper;
+import com.magenta.security.SecurityService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -20,11 +28,12 @@ import java.util.List;
 @Transactional
 public class UserServiceImpl implements UserService {
 
+    private static final Logger LOGGER = LogManager.getLogger(UserServiceImpl.class);
+
     private final UserDao userDao;
     private final ContractService contractService;
-    private final ModelMapper modelMapper;
-
-
+    private final MyModelMapper modelMapper;
+    private final SecurityService securityService;
     private PasswordEncoder passwordEncoder;
 
     @Autowired
@@ -33,10 +42,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Autowired
-    public UserServiceImpl(UserDao userDao, ModelMapper modelMapper, ContractService contractService){
+    public UserServiceImpl(UserDao userDao, MyModelMapper modelMapper, ContractService contractService,
+                           SecurityService securityService){
         this.userDao = userDao;
         this.modelMapper = modelMapper;
         this.contractService = contractService;
+        this.securityService = securityService;
     }
 
     @Override
@@ -45,11 +56,12 @@ public class UserServiceImpl implements UserService {
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         user.setStatus(Status.ACTIVE);
         user.setRole(Role.USER);
+        user.setBalance(0.0);
         userDao.save(user);
     }
 
     @Override
-    public UserDto findById(int id) throws MyException {
+    public UserDto findById(int id) throws DatabaseException {
         return modelMapper.map(userDao.findById(id),UserDto.class);
     }
 
@@ -61,27 +73,61 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserDto findByNumber(String number) throws MyException {
-        ContractDto contractDto = contractService.findByNumber(number);
+    public UserDto findByNumberOrEmail(String request) throws DatabaseException {
+
+        if(request.contains("@")){
+            User user = userDao.findByEmail(request);
+            return modelMapper.map(user, UserDto.class);
+        }
+        ContractDto contractDto = contractService.findByNumber(request);
         return modelMapper.map(contractDto.getUser(),UserDto.class);
     }
 
     @Override
-    public User findByLogin(String login) throws MyException {
-        return userDao.findByLogin(login);
+    public UserDto findByLogin(String login) throws DatabaseException {
+        return modelMapper.map(userDao.findByLogin(login), UserDto.class);
     }
 
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id) throws DatabaseException, MyException {
+        if (!userDao.findById(id).getNumbers().isEmpty()){
+            throw new MyException("User still have one or more contracts.");
+        }
         userDao.delete(id);
     }
 
 
     @Override
-    public void update(UserDto userDto) {
+    public void update(EditUserDto editDto) {
+        UserDto userDto = modelMapper.map(editDto, UserDto.class);
         User user = modelMapper.map(userDto,User.class);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         userDao.update(user);
+    }
+
+    @Override
+    public void setStatus(ChangeStatusDto statusDto) throws AuthorizationException, DatabaseException {
+
+        User user = userDao.findById(statusDto.getEntityId());
+        
+        if (!securityService.isAdmin() && user.getStatus().equals(Status.BLOCKED)){
+            throw new AuthorizationException("Only administrator can do it");
+        }
+
+        Status newStatus = modelMapper.map(statusDto.getStatus(),Status.class);
+        user.setStatus(newStatus);
+    }
+
+    @Override
+    public void addFunds(AddFundsDto funds) throws DatabaseException, MyException {
+        User user = userDao.findById(funds.getUserId());
+
+        if (!user.getStatus().equals(Status.ACTIVE)){
+            throw new MyException("User isn't active!");
+        }
+        double old = user.getBalance();
+        double add = funds.getFunds();
+        user.setBalance(old + add);
     }
 }
